# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UNIVGO API Documentation",
        "description": "API Documentation for UNIVGO application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/news/{news}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get news detail by ID",
                "operationId": "4c09fa11f86d11aa99aefdc9f85c1237",
                "parameters": [
                    {
                        "name": "news",
                        "in": "path",
                        "description": "News ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "News detail and related news"
                    },
                    "404": {
                        "description": "News not found"
                    }
                }
            }
        },
...
    },
...
}